+++ packages/plugin-nodemailer/src/plugin.ts
@@ -1,20 +1,22 @@
 import T from "./translations/index.js";
 import verifyTransporter from "./utils/verify-transporter.js";
+import isValidData from "./utils/is-valid-data.js";
+import { PLUGIN_KEY, LUCID_VERSION } from "./constants.js";
 import type { LucidPluginOptions } from "@lucidcms/core/types";
 import type { PluginOptions } from "./types/types.js";
-import { PLUGIN_KEY, LUCID_VERSION } from "./constants.js";
 
 const plugin: LucidPluginOptions<PluginOptions> = async (
 	config,
 	pluginOptions,
 ) => {
 	config.email = {
+		identifier: "nodemailer",
 		from: pluginOptions.from,
 		strategy: async (email, meta) => {
 			try {
 				await verifyTransporter(pluginOptions.transporter);
 
-				await pluginOptions.transporter.sendMail({
+				const data = await pluginOptions.transporter.sendMail({
 					from: `${email.from.name} <${email.from.email}>`,
 					to: email.to,
 					subject: email.subject,
@@ -28,12 +30,13 @@ const plugin: LucidPluginOptions<PluginOptions> = async (
 				return {
 					success: true,
 					message: T("email_successfully_sent"),
+					data: isValidData(data) ? data : null,
 				};
 			} catch (error) {
-				const err = error as Error;
 				return {
 					success: false,
-					message: err.message,
+					message:
+						error instanceof Error ? error.message : T("email_failed_to_send"),
 				};
 			}
 		},